requirements.txt
astroid==2.15.5
backcall==0.2.0
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
dill==0.3.6
executing==1.2.0
fonttools==4.40.0
greenlet==2.0.2
haversine==2.8.0
isort==5.12.0
joblib==1.2.0
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
matplotlib==3.7.1
mccabe==0.7.0
multimethod==1.9.1
mypy-extensions==1.0.0
numpy==1.24.3
packaging==23.1
pandas==2.0.2
pandera==0.15.1
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.5.1
psutil==5.9.5
psycopg2==2.9.6
pure-eval==0.2.2
pyarrow==12.0.0
pydantic==1.10.8
Pygments==2.15.1
pylint==2.17.4
pyparsing==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pywin32==306
PyYAML==6.0
pyzmq==25.1.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sklearn==0.0.post5
SQLAlchemy==2.0.16
threadpoolctl==3.1.0
tomli==2.0.1
tomlkit==0.11.8
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typeguard==4.0.0
typing-inspect==0.9.0
typing_extensions==4.6.2
tzdata==2023.3
wcwidth==0.2.6
wrapt==1.15.0
